Orginal file from: https://social.technet.microsoft.com/Forums/en-US/9ae52a5d-68cb-420b-b409-9c7817d8eb67/windows-8-activation-through-mdt Edited by aghoneim Friday, March 08, 2013 11 : 12 PM moved code to the end of the reply.
You could actually get the key from ACPI tables.The Windows API has two functions :

EnumSystemFirmwareTables
GetSystemFirmwareTable

that you can use to read the tables from BIOS.The product key is stored in the Microsoft Software Licensing Tables(SLIC and MSDM)

Here is some code that would do that, it's a quick hack but it spits the key out to use in a batch file with slmgr, or supply -v to display more info. I tested it with a couple of models and word with no problems.

Build with VC++ as a statically linked exe.